requirements.txt
Django==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.1.0
qrcode==7.4.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
celery==5.3.4
